requirements.txt
Django==4.2.8
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
daphne==4.0.0
